+++ CHANGELOG.md
@@ -4,6 +4,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased][unreleased]
  * Add the `language-*` class also to the `<pre>` tag (not only to `<code>`).
+ * Add the class `language-x` if `x` is the defined language, even if `x` cannot be highlighted with Prism.
 
 ## [1.0.1] 2016-10-19
  * Fixed build bug that led to the plugin not being exported correctly.